+++ packages/react/CHANGELOG.md
@@ -1,4 +1,4 @@
-## Next
+## 1.0.0-beta.14
 
 * Make passing the `action` to `<CuriBase>` optional (it will default to `POP`).
 * Rename `<Navigator>` to `<CuriBase>`.